streamingConsumer/requirements.txt
confluent_kafka
kafka-python
psycopg2
